RSC: Extract fetchRsc out into its own file
